开始的增删改查

+++ src/main/java/com/itheima/mp/domain/vo/CourseVO.java
@@ -11,7 +11,7 @@ public class CourseVO {
     private Integer courseId;
     private String num;
     private String name;
-    private Integer credit;
+    private double credit;
 
     private double score;
     private Integer rankClass;
